Add basic project structure to JSlim
Going forward the project will be much easier to add, remove, or make updates to without the entire logic for the program living in index.js
