+++ src/index.js
@@ -1,0 +1,3 @@
+import JSlim from "./plugin/JSlim";
+
+JSlim();
